+++ CHANGES.md
@@ -1,3 +1,10 @@
+# Version 0.4.0.0 (Unknown)
+
+  * Add the `resolve` function.  If you have a playlist that contains
+    references to remote playlists, this new function will recursively
+    download and process playlists until it has a flat playlist with
+    no further remote references.
+
 # Version 0.3.0.0 (January 24, 2015)
 
   * Update dependencies to recent versions (thanks to Wieland Hoffmann).